Started Colonist Collection System
Imported skybox, "Starfield Skybox," from the Unity asset store and implemented it. More importantly, I've started building the game's colonist collection system. I've created a colonist prefab and filled my prototype scene with several colonists and started creating the UI. I then created some scripts that detect when the player collides with a colonist, disables the colonist object, increment the colonist counter in the player's inventory, and display that information in the UI on the screen. This was all done using the following videos as references:
- https://www.youtube.com/watch?v=EfUCEwKmcjc
- https://www.youtube.com/watch?v=3uyolYVsiWc