MedicalEmergency Scriptable Object
I created the "MedicalEmergency" scriptable object. From what I understand, each instance of the "MedicalEmergency" scriptable object represents a medical emergency a drifting colonist can have.  Each "MedicalEmergency" scriptable object stores data regarding the medical emergency it represents, such as the name of the emergency, a list of regions of the body where this medical emergency is localized, and how much time it will take to stabilize it. Additionally, the drifting colonist prefab now has the long-forgotten "Colonist" script attached to it which contains a list of medical emergencies the colonist is experiencing.