New methodology for collecting colonists
Previously, when the player collides with a colonist game object, the "PlayerInventory" script attached to the player game object used an "OnCollisionEnter" method which checked whether the object had the "Colonist" tag and whether there was enough room to carry another colonist. If both are true, the "CollonistCollected" method, which increases the "colonistNum" integer variable by one, is called and the colonist game object the player collided with gets deactivated.

This is now no longer the case. In this new implementation, the "PlayerInventory" has a private list for type colonist called "storedColonists" in which the information of collected colonists will be stored and the previous implementations of the "colonistNum" variable have been replaced with the use of "storedColonists.Count" which is the total count of colonists in "storedColonists". The "CollonistCollected" method has been renamed the "CollectColonist" method which takes in a type Colonist variable "colonist". It first checks whether "storedColonists.Count" is less than or equal to the number of slots available, "slotNum", if it is not,  "colonist" will be added to "storedColonists" before the colonist in the scene that "colonist" represents is deactivated. Moving on, the "Colonist" script attached to the colonist game object uses an "OnCollisionEnter" method to first check whether what it collided with had the "Player" tag and then whether the object had a "PlayerInventory" script attached to it. If both these conditions are met, it calls on the object's "PlayerInventory" script and calls the "CollectColonist" method using itself as the input variable, "playerInventory.CollectColonist(this)". These are all the changes that were made, aside from that, I learned how to use formated strings with Debug.Log, "Debug.Log($"Colonist Collected. Slots remaining: {slotNum - storedColonists.Count}/{slotNum}");".