Eject Colonists + UpdateUISlots()
Now, when the player hits the "X" key it triggers the eject action which, for now, ejects the colonist stored at the first location in the "storedColonists" list 1 unit in front of the player at a velocity moving away from the player along its local z-axis. Additionally, I made the "PlayerInventory" script more efficient, it no longer updates the colonist slots in the UI every frame, now a "UpdateUISlots" method updates the slots whenever it is called. The "UpdateUISlots" method is first called at "Start()" as an initial update and it is also called within the " CollectColonist" and "EjectColonist" methods because that's when the UI slots would need to be updated.