Sight Range and Attack Range Gizmos Fixed
For one, I needed to call transform directly instead of "agentLocation" since "OnDrawGizmosSelected" is called before the "Start" method which is where "agentLocation" is set. Secondly, the reason the ranges wouldn't appear unless the enemy is selected was because I was using the "OnDrawGizmosSelected" method when I should have been using "OnDrawGizmos" method so all ranges show. Unfortunately, when I used this it caused the game to slow down so I'll be sticking with the "OnDrawGizmosSelected" method.