Cycle through stored colonists
The player can now cycle back and forth between colonist slots using the "Q" (previous) and "E" (next) keys. However, it can only cycle through the slots that are filled and it is still the case that it is the first colonist in the "storedColonist" list that gets ejected when "EjectColonist" is called.