Colonist Selection System Working as Expected
The "storedColonists" list has been renamed as "slotList" which, after the new "InstantiateSlotList()" method has been called, contains a number of nulls equivalent to "slotNum" with each null representing an empty slot and a new int "filledSlots" keeping count of the number of filled slots in the list. Additionally, the "UpdateUISlots()" method has been updated, now an empty slot is rendered if an index in "slotList" is equal to null and is rendered as a full slot if it is not.

Whenever the "CollectColonist(Colonist colonist)" method is called it will be first checked "filledSlots >= slotNum" to see whether there's room for a colonist. Next, after increasing "filledSlots" by 1, if the slot at the selected index is empty, "if (slotList[selectedColonistIndex] == null)", the "slotList" index that is the selected index will be set equal to the collected colonist, otherwise it will do so with the first empty slot, aka the index of the first index in "slotList" that equals null will be set equal to the collected colonist. Finally, the gameObject that the collected colonist was attached to gets deactivated and "UpdateUISlots()" is called to update the slot UI.

The "EjectColonist(Colonist colonist)" method has been replaced with "EjectColonist()" which does not take in an input colonist variable. Unless the selected slot index of "slotList" is equal to null, "if (slotList[selectedColonistIndex] != null)",  "filledSlots" is first decreased by 1 and the selected slot index of "slotList" is set equal to null. Next the transform of the selected colonist is set to be one unit in front of the player game object and then the selected colonist game object is reactivated. Finally, after "UpdateUISlots()" is called, the rigidbody's velocity attached to the selected colonist game object is set to be 1 in the direction forward of the player game object.