Fixed CPB bug and added to Emergency/Region prefabs
I found out how to fix a bug I was experiencing with the CircularProgressionBarEditor script and after that I added Circular Progress Bar prefabs to both the Emergency and the Region prefabs. Now all I have to do is make it so that the progress bars fill when the player heals a region and have each emergency's progress bar reflect the progress of all its region progresses.