Healing Ore objects now heal the player immediately
I added "Heal(float healAmount)" to the "PlayerHealth" script and modified "OnCollisionEnter(Collision collision)" so that when the player collides with a Healing Ore object, the player gains more health, and the healing object is destroyed. I also modified "Drag()" in the "PlayerController" script so that it can handle when a game object is destroyed while the player is still holding it.