Breakable Asteroids Drop Health Ore
I first downloaded the free asset "Breakable Asteroids" from the Unity asset store which I found to be a very useful addition to my project. Additionally, the reviews on the asset's download page helped me understand how to code for the asset in my project:
- https://assetstore.unity.com/packages/3d/props/breakable-asteroids-167825#reviews

Using the asset as well as a script I made "BreakableAsteroid" I was able to create breakable asteroids that, when slammed into walls, will break if the force of the collision exceeded "fractureForce" and the object was still being held by the player when it collided. I used the following page as reference for figuring out how to find the collision force:
- https://discussions.unity.com/t/what-force-did-collider-collide-with/33918
Additionally, "BreakableAsteroid" also spawns a lump of healing ore when its asteroid is fractured for the player. When an asteroid fractures, the "BreakableAsteroid" instantiates the fractured version of the asteroid before destroying the gameObject, this fractured object has a "FadePieces" script attached to it that will fade and destroy the pieces after a certain amount of time.

After that, I attempted to push this commit to origin but I got the warning that "zasteroids.dae" was over 100MB and could not be pushed. Before I tried fixing this, I tried getting started with the system where collected Health ore is stored and used by pressing "F" however there is an error with it that I am trying to figure that out.

Finally, I tried fixing the problem preventing me from pushing this commit. I discovered that GitHub Desktop has an "Open in Command Prompt" option listed under "Repository" and from there I used the following references to figure out how to use Git LFS "zasteroids.dae" so this commit could be pushed:
- https://docs.github.com/en/repositories/working-with-files/managing-large-files/configuring-git-large-file-storage
- https://docs.github.com/en/repositories/working-with-files/managing-large-files/moving-a-file-in-your-repository-to-git-large-file-storage
- https://github.com/desktop/desktop/issues/5217
- https://github.com/desktop/desktop/issues/997
- https://stackoverflow.com/questions/65820300/pushing-files-over-100mb-to-github
- https://stackoverflow.com/questions/5989592/git-cannot-checkout-branch-error-pathspec-did-not-match-any-files-kn