Colonist Ragdolls
After importing the updated third version of "Colonist_Alpha", which is 1.5 times the size of the previous version, I created a new scene called "Colonist Ragdoll Test Lab" to create a colonist rag doll. I set the layer of each limb with a collider to "Grabable", unchecked gravity for all the rigidbodies present in the ragdoll, and tweaked the size of some of the colliders. Then, finally, I checked "Y" for "Freeze Position" in the "spine1" rigidbody. It seems to be working better than I expected. The only bug is that when moved back and forth extremely quickly, the limbs get very stretched. As a final note, I also created a "Colonist Animator Controller".